+++ audionimbus/src/model/air_absorption.rs
@@ -2,7 +2,7 @@ use crate::context::Context;
 use crate::geometry;
 
 /// An air absorption model that can be used for modeling frequency-dependent attenuation of sound over distance.
-#[derive(Debug)]
+#[derive(Debug, Copy, Clone)]
 pub enum AirAbsorptionModel {
     /// The default air absorption model.
     /// This is an exponential falloff, with decay rates derived from physical properties of air.