+++ audionimbus/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Implement `Send` trait for `Scene`.
+- Refactor `SimulationInputs`.
 
 ## [0.2.0] - 2025-03-17